requirements.txt
anthropic==0.3.11
fastapi==0.104.1
gradio==4.8.0
redis==5.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
paramiko==3.3.1
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
